Update lesson objectives
This is my first honest attempt at writing a good commit
message.
Focusing on the what changed and why:
-what changed is the content of README.md file
-in order to state the objectives for doing this lesson;
after finishing this, will be useful to come back and reflect
on whether these objectives were achieved, and what else
was learned and what could have been improved in the
README.md file.